+++ packages/ts-components/src/components/banner/assets/CloseIconBlack.tsx
