+++ packages/ts-components/src/components/banner/assets/EmailIcon.tsx
