+++ packages/ts-newskit/src/assets/index.tsx
@@ -1,5 +1,3 @@
 export { default as TooltipIcon } from '../assets/TooltipIcon';
-export { default as CloseIconBlack } from '../assets/CloseIconBlack';
-export { default as EmailIcon } from '../assets/EmailIcon';
 export { default as FilledArrowIcon } from './FilledArrowIcon';
 export { default as BreadcrumbIcon } from './BreadcrumbIcon';
