+++ packages/ts-newskit/src/index.ts
@@ -6,6 +6,5 @@ export { JobTitle } from './components/misc/job-title/job-title';
 export {
   UpdateButtonWithDelay
 } from './components/misc/update-button/update-button-with-delay';
-export { Banner } from './components/misc/banner/banner';
 
 // TODO Remove after tests - borisovy
